Move type creation code to separate module
