+++ apps/web/modules/type/index.ts
@@ -1,0 +1,1 @@
+export * as CreateType from './create-type';
